# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Status": {
				"enum": [
					"true",
					"false"
				],
				"type": "string"
			},
			"BaseResponse": {
				"properties": {
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"required": [
					"description",
					"status"
				],
				"type": "object"
			},
			"Category": {
				"properties": {
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"id"
				],
				"type": "object"
			},
			"CategoriesResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"feeds": {
								"items": {
									"$ref": "#/components/schemas/Category"
								},
								"type": "array"
							},
							"count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"feeds",
							"count"
						],
						"type": "object"
					}
				]
			},
			"Stats": {
				"properties": {
					"feedsWithValueBlocks": {
						"type": "number",
						"format": "double"
					},
					"feedsWithNewEpisodes90days": {
						"type": "number",
						"format": "double"
					},
					"feedsWithNewEpisodes30days": {
						"type": "number",
						"format": "double"
					},
					"feedsWithNewEpisodes10days": {
						"type": "number",
						"format": "double"
					},
					"feedsWithNewEpisodes3days": {
						"type": "number",
						"format": "double"
					},
					"episodeCountTotal": {
						"type": "number",
						"format": "double"
					},
					"feedCountTotal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"feedsWithValueBlocks",
					"feedsWithNewEpisodes90days",
					"feedsWithNewEpisodes30days",
					"feedsWithNewEpisodes10days",
					"feedsWithNewEpisodes3days",
					"episodeCountTotal",
					"feedCountTotal"
				],
				"type": "object"
			},
			"StatsResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"stats": {
								"$ref": "#/components/schemas/Stats"
							}
						},
						"required": [
							"stats"
						],
						"type": "object"
					}
				]
			},
			"Record_number.string_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Feed": {
				"properties": {
					"categories": {
						"$ref": "#/components/schemas/Record_number.string_"
					},
					"language": {
						"type": "string"
					},
					"trendScore": {
						"type": "number",
						"format": "double"
					},
					"itunesId": {
						"type": "number",
						"format": "double"
					},
					"newestItemPublishTime": {
						"type": "number",
						"format": "double"
					},
					"artwork": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"categories",
					"language",
					"trendScore",
					"itunesId",
					"newestItemPublishTime",
					"artwork",
					"image",
					"author",
					"description",
					"title",
					"url",
					"id"
				],
				"type": "object"
			},
			"TrendingResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"since": {
								"type": "number",
								"format": "double"
							},
							"max": {
								"type": "number",
								"format": "double"
							},
							"count": {
								"type": "number",
								"format": "double"
							},
							"feeds": {
								"items": {
									"$ref": "#/components/schemas/Feed"
								},
								"type": "array"
							}
						},
						"required": [
							"since",
							"max",
							"count",
							"feeds"
						],
						"type": "object"
					}
				]
			},
			"RecentEpisodeFeed": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Feed"
					},
					{
						"properties": {
							"feedLanguage": {
								"type": "string"
							},
							"feedTitle": {
								"type": "string"
							},
							"feedId": {
								"type": "number",
								"format": "double"
							},
							"feedImage": {
								"type": "string"
							},
							"feedItunesId": {
								"type": "number",
								"format": "double"
							},
							"season": {
								"type": "number",
								"format": "double"
							},
							"episodeType": {
								"type": "string"
							},
							"episode": {
								"type": "number",
								"format": "double"
							},
							"explicit": {
								"type": "number",
								"format": "double"
							},
							"enclosureLength": {
								"type": "number",
								"format": "double"
							},
							"enclosureType": {
								"type": "string"
							},
							"enclosureUrl": {
								"type": "string"
							},
							"dateCrawled": {
								"type": "number",
								"format": "double"
							},
							"datePublished": {
								"type": "number",
								"format": "double"
							},
							"guid": {
								"type": "string"
							},
							"link": {
								"type": "string"
							}
						},
						"required": [
							"feedLanguage",
							"feedTitle",
							"feedId",
							"feedImage",
							"feedItunesId",
							"season",
							"episodeType",
							"episode",
							"explicit",
							"enclosureLength",
							"enclosureType",
							"enclosureUrl",
							"dateCrawled",
							"datePublished",
							"guid",
							"link"
						],
						"type": "object"
					}
				]
			},
			"RecentEpisodesResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"max": {
								"type": "number",
								"format": "double"
							},
							"count": {
								"type": "number",
								"format": "double"
							},
							"items": {
								"items": {
									"$ref": "#/components/schemas/RecentEpisodeFeed"
								},
								"type": "array"
							}
						},
						"required": [
							"max",
							"count",
							"items"
						],
						"type": "object"
					}
				]
			},
			"Item": {
				"properties": {
					"feedId": {
						"type": "number",
						"format": "double"
					},
					"feedUrl": {
						"type": "number",
						"format": "double"
					},
					"feedTitle": {
						"type": "number",
						"format": "double"
					},
					"episodeTitle": {
						"type": "number",
						"format": "double"
					},
					"episodeId": {
						"type": "number",
						"format": "double"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"enclosureUrl": {
						"type": "string"
					}
				},
				"required": [
					"feedId",
					"feedUrl",
					"feedTitle",
					"episodeTitle",
					"episodeId",
					"duration",
					"startTime",
					"title",
					"enclosureUrl"
				],
				"type": "object"
			},
			"SoundbitesResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"items": {
								"items": {
									"$ref": "#/components/schemas/Item"
								},
								"type": "array"
							}
						},
						"required": [
							"items"
						],
						"type": "object"
					}
				]
			},
			"SearchFeed": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Feed"
					},
					{
						"properties": {
							"newestItemPubdate": {
								"type": "number",
								"format": "double"
							},
							"imageUrlHash": {
								"type": "number",
								"format": "double"
							},
							"locked": {
								"type": "number",
								"format": "double"
							},
							"parseErrors": {
								"type": "number",
								"format": "double"
							},
							"crawlErrors": {
								"type": "number",
								"format": "double"
							},
							"episodeCount": {
								"type": "number",
								"format": "double"
							},
							"dead": {
								"type": "number",
								"format": "double"
							},
							"type": {
								"type": "number",
								"format": "double"
							},
							"explicit": {
								"type": "boolean"
							},
							"generator": {
								"type": "string"
							},
							"contentType": {
								"type": "string"
							},
							"lastHttpStatus": {
								"type": "number",
								"format": "double"
							},
							"lastGoodHttpStatusTime": {
								"type": "number",
								"format": "double"
							},
							"lastParseTime": {
								"type": "number",
								"format": "double"
							},
							"lastCrawlTime": {
								"type": "number",
								"format": "double"
							},
							"lastUpdateTime": {
								"type": "number",
								"format": "double"
							},
							"ownerName": {
								"type": "string"
							},
							"link": {
								"type": "string"
							},
							"originalUrl": {
								"type": "string"
							},
							"podcastGuid": {
								"type": "string"
							}
						},
						"required": [
							"newestItemPubdate",
							"imageUrlHash",
							"locked",
							"parseErrors",
							"crawlErrors",
							"episodeCount",
							"dead",
							"type",
							"explicit",
							"generator",
							"contentType",
							"lastHttpStatus",
							"lastGoodHttpStatusTime",
							"lastParseTime",
							"lastCrawlTime",
							"lastUpdateTime",
							"ownerName",
							"link",
							"originalUrl",
							"podcastGuid"
						],
						"type": "object"
					}
				]
			},
			"SearchResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BaseResponse"
					},
					{
						"properties": {
							"query": {
								"type": "string"
							},
							"count": {
								"type": "number",
								"format": "double"
							},
							"feeds": {
								"items": {
									"$ref": "#/components/schemas/SearchFeed"
								},
								"type": "array"
							}
						},
						"required": [
							"query",
							"count",
							"feeds"
						],
						"type": "object"
					}
				]
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "podcasts-api",
		"version": "1.0.0",
		"description": "A wrapper api for podcast index",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Krishna Mohan A M"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoriesResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/stats": {
			"get": {
				"operationId": "GetStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/trending": {
			"get": {
				"operationId": "GetTrending",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrendingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "lang",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cat",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recentEpisodes": {
			"get": {
				"operationId": "GetRecentEpisodes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecentEpisodesResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "before",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/podcastsFeed": {
			"get": {
				"operationId": "GetPodcastsFeed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrendingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/episodesFeed": {
			"get": {
				"operationId": "GetEpisodesFeed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrendingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/soundbites": {
			"get": {
				"operationId": "GetSoundbites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SoundbitesResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/search": {
			"get": {
				"operationId": "SearchByTerm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "clean",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "fulltext",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}